Dind volume path filter in ./utils
[skip ci]
