+++ src/external-server/external-server.ts
@@ -43,28 +43,6 @@ export async function getFreePort(): Promise<string> {
     throw new StarknetPluginError("Could not find a free port, try rerunning your command!");
 }
 
-/**
- * Returns function to filter host path for docker volumes.
- * Uses environment variable STARKNET_HARDHAT_DIND_HOST_PATH
- * The variable is set when running from inside DinD container
- * DinD = Docker in Docker
- * @returns {function} Callback to fix docker volume host path
- */
-export function getVolumeHostPathFilter(): (path: string) => string {
-    // If environment variable is set to replace docker host path
-    const { STARKNET_HARDHAT_DIND_HOST_PATH = "" } = process.env;
-
-    if (STARKNET_HARDHAT_DIND_HOST_PATH && STARKNET_HARDHAT_DIND_HOST_PATH.includes(":")) {
-        // We have paths to replace
-        const replacement = STARKNET_HARDHAT_DIND_HOST_PATH.split(":");
-        if (replacement[0] && replacement[1])
-            // Return volumeHostPathFilter function
-            return (path: string) => path.replace(replacement[0], replacement[1]);
-    }
-    // Returns a dummy function
-    return (path: string) => path;
-}
-
 export abstract class ExternalServer {
     protected childProcess: ChildProcess;
     private connected = false;
