+++ src/utils.ts
@@ -78,6 +78,28 @@ export function adaptUrl(url: string): string {
     return url;
 }
 
+/**
+ * Returns function to filter host path for docker volumes.
+ * Uses environment variable STARKNET_HARDHAT_DIND_HOST_PATH
+ * The variable is set when running from inside DinD container
+ * DinD = Docker in Docker
+ * @returns {function} Callback to fix docker volume host path
+ */
+export function getDindVolumeHostPathFilter(): (path: string) => string {
+    // If environment variable is set to replace docker host path
+    const { STARKNET_HARDHAT_DIND_HOST_PATH = "" } = process.env;
+
+    if (STARKNET_HARDHAT_DIND_HOST_PATH && STARKNET_HARDHAT_DIND_HOST_PATH.includes(":")) {
+        // We have paths to replace
+        const replacement = STARKNET_HARDHAT_DIND_HOST_PATH.split(":");
+        if (replacement[0] && replacement[1])
+            // Return volumeHostPathFilter function
+            return (path: string) => path.replace(replacement[0], replacement[1]);
+    }
+    // Returns a dummy function
+    return (path: string) => path;
+}
+
 export function getDefaultHttpNetworkConfig(
     url: string,
     verificationUrl: string,
